Commentaire limité à 1 signalement et possibilité d'ignorer une modération

+++ view/verification/moderation.php
@@ -15,15 +15,53 @@ if(isset($_POST['addModeration'])){
 
 			$moderationManager = new Arthur\WriterBlog\Model\ModerationEntityManager();
 
-			$addModeration = $moderationManager->addModeration($moderation);
-			if ($addModeration === false) {
-				echo '<meta http-equiv="refresh" content="0;URL=index.php?action=post&id='.$postId.'&addModeration=no">';
-	        }
-	        else {
-	        	echo '<meta http-equiv="refresh" content="0;URL=index.php?action=post&id='.$postId.'&addModeration=yes">';
-	        }
+			$existModeration = $moderationManager->getModerationComment($moderation);
+
+			if(!empty($existModeration)){
+				$addModeration = $moderationManager->addModeration($moderation);
+				if ($addModeration === false) {
+					echo '<meta http-equiv="refresh" content="0;URL=index.php?action=post&id='.$postId.'&addModeration=no">';
+		        }
+		        else {
+		        	echo '<meta http-equiv="refresh" content="0;URL=index.php?action=post&id='.$postId.'&addModeration=yes">';
+		        }
+				
+				
+			}
+			else{
+				echo '<meta http-equiv="refresh" content="0;URL=index.php?action=post&id='.$postId.'&moderation=exist">';
+		    }
 		}
 	}
+}
+
+if(isset($_POST['ignoreModeration'])){
+	if(isset($_POST['moderationId']) && !empty(trim($_POST['moderationId']))){
+		extract($_POST);
+
+		$moderationId = $_POST['moderationId'];
+
+		$moderation = new ModerationEntity();
+		$moderation->setId($moderationId);
+
+		$moderationManager = new Arthur\WriterBlog\Model\ModerationEntityManager(); //Test si billet de modération existe
+		$existModeration = $moderationManager->getModeration($moderation);
+
+		if(!empty($existModeration)){
+			$deleteModeration = $moderationManager->deleteModeration($moderation);
+
+			if ($deleteModeration === false) {
+	    		echo '<meta http-equiv="refresh" content="0;URL=index.php?action=moderation&deleteModeration=no">';  
+	        }
+	        else {
+	        	echo '<meta http-equiv="refresh" content="0;URL=index.php?action=moderation&ignoreModeration=yes">';
+	        }
+		}
+		else{
+			echo '<meta http-equiv="refresh" content="0;URL=index.php?moderation&existModeration=no">';  
+		}
+	}
+	
 }
 
 if(isset($_POST['updateModeration'])){
@@ -60,7 +98,7 @@ if(isset($_POST['updateModeration'])){
 			    }
 			    else {
 			    	$deleteModeration = $moderationManager->deleteModeration($moderation); //Suppression du billet de modération
-			    	if ($updateComment === false) {
+			    	if ($deleteModeration === false) {
 			    		echo '<meta http-equiv="refresh" content="0;URL=index.php?action=moderation&deleteModeration=no">';  
 			        }
 			        else {
